WIP test: make 96% of tests pass on single endpoint test suite
